amba_sys2elf: Implemented reading of base address.
Also made changing entry point within the ELF.

+++ amba_sys2elf.py
@@ -19,11 +19,13 @@ def eprint(*args, **kwargs):
 
 class ProgOptions:
   fwpartfile = ''
+  basename = ''
   outfile = ''
   inifile = ''
   section_pos = { '.text': 0x0, '.ARM.exidx': -2, '.dsp_buf': -3, '.data': -4, 'no_init': -5, '.bss.noinit': -6, '.bss': -7 }
   elftemplate='amba_sys2elf_template.elf'
   verbose = 0
+  dry_run = False
   command = ''
 
 # Converts "System Software" partition from Ambarella a7/a9 firmware
@@ -77,12 +79,31 @@ class ProgOptions:
 def section_is_bss(sectname):
   return sectname.startswith('.bss')
 
+def syssw_read_base_address(po):
+  mem_addr = 0
+  fname = "{:s}.a9h".format(po.basename)
+  if (po.verbose > 1):
+    print("{}: Opening {:s}".format(po.fwpartfile,fname))
+  parser = configparser.ConfigParser()
+  with open(fname, "r") as lines:
+    lines = itertools.chain(("[asection]",), lines) # This line adds section header to ini
+    parser.read_file(lines)
+    mem_addr = int(parser.get("asection", "mem_addr"),16)
+  del parser
+  return mem_addr
+
 def syssw_bin2elf(po, fwpartfile):
-  # TODO read base address
-  memaddr_base = 0xa0001000
-  # TODO detect offsets
+  # read base address from INI file which should be there after AMBA extraction
+  memaddr_base = syssw_read_base_address(po)
+  if (po.verbose > 0):
+    print("{}: Memory base address set to 0x{:08x}".format(po.fwpartfile,memaddr_base))
+  # detect position of each section in the binary file
+  if (po.verbose > 1):
+    print("{}: Searching for sections".format(po.fwpartfile))
+  # TODO
   # set positions for bss sections too - to the place
   # where they yould be if they existed
+  # TODO
   # Prepare list of sections in the order of position
   sections_order = []
   for sortpos in sorted(po.section_pos.values()):
@@ -101,18 +122,26 @@ def syssw_bin2elf(po, fwpartfile):
   for sectname in sections_order:
     if section_is_bss(sectname):
       continue
-    fname = "{:s}_sect_{:s}.bin".format(os.path.basename(po.fwpartfile),re.sub('[\W_]+', '', sectname))
+    if (po.verbose > 0):
+      print("{}: Extracting section '{:s}' from binary pos 0x{:08x}".format(po.fwpartfile,sectname,po.section_pos[sectname]))
+    fname = "{:s}_sect_{:s}.bin".format(po.basename, re.sub('[\W_]+', '', sectname))
     sections_fname[sectname] = fname
-    sectfile = open(fname, "wb")
+    if not po.dry_run:
+      sectfile = open(fname, "wb")
     fwpartfile.seek(po.section_pos[sectname], os.SEEK_SET)
     n = 0
     while (n < sections_size[sectname]):
       copy_buffer = fwpartfile.read(min(1024 * 1024, sections_size[sectname] - n))
       if not copy_buffer:
+          raise EOFError("Couldn't read whole section '{:s}' from binary file.".format(sectname))
           break
       n += len(copy_buffer)
-      sectfile.write(copy_buffer)
-    sectfile.close()
+      if not po.dry_run:
+        sectfile.write(copy_buffer)
+    if not po.dry_run:
+      sectfile.close()
+    if (po.verbose > 1):
+      print("{}: Section '{:s}' written to '{:s}'".format(po.fwpartfile,sectname,fname))
   # Prepare array of addresses
   sections_address = {}
   memaddr = memaddr_base
@@ -123,6 +152,8 @@ def syssw_bin2elf(po, fwpartfile):
     # add section address to array
     sections_address[sectname] = memaddr
     memaddr += sections_size[sectname]
+    if (po.verbose > 0):
+      print("{}: Section '{:s}' memory address set to 0x{:08x}".format(po.fwpartfile,sectname,sections_address[sectname]))
   # prepare objcopy command line
   objcopy_cmd = '/usr/bin/arm-none-eabi-objcopy'
   for sectname in sections_order:
@@ -131,14 +162,24 @@ def syssw_bin2elf(po, fwpartfile):
     objcopy_cmd += ' --change-section-address "{0:s}=0x{1:08x}"'.format(sectname,sections_address[sectname])
   objcopy_cmd += ' "{:s}" "{:s}"'.format(po.elftemplate, po.outfile)
   # execute the objcopy command
-  objcopy_cmd = "echo "+objcopy_cmd
-  retcode = subprocess.call(objcopy_cmd, shell=True)
+  if (po.verbose > 0):
+    print("{}: Executing command:\n{:s}".format(po.fwpartfile,objcopy_cmd))
+  retcode = 0
+  if not po.dry_run:
+    retcode = subprocess.call(objcopy_cmd, shell=True)
+  if (retcode != 0):
+    raise EnvironmentError("Execution of objcopy returned with error {:d}.".format(retcode))
+  # update entry point in the ELF header
+  elffile = open(po.outfile, "r+b")
+  elffile.seek(16+2*4, os.SEEK_SET) # position of e_entry within ELF header
+  elffile.write(c_uint(memaddr_base))
+  elffile.close()
 
 def main(argv):
   # Parse command line options
   po = ProgOptions()
   try:
-     opts, args = getopt.getopt(argv,"hevt:p:s:o:",["help","version","mkelf","fwpart=","template","section=","output="])
+     opts, args = getopt.getopt(argv,"hevt:p:s:o:",["help","version","mkelf","dry-run","fwpart=","template","section=","output="])
   except getopt.GetoptError:
      print("Unrecognized options; check amba_sys2elf.py --help")
      sys.exit(2)
@@ -153,8 +194,10 @@ def main(argv):
      elif opt == "--version":
         print("amba_sys2elf.py version 0.1.0")
         sys.exit()
-     elif opt == '-v':
+     elif opt == "-v":
         po.verbose += 1
+     elif opt == "--dry-run":
+        po.dry_run = True
      elif opt in ("-p", "--fwpart"):
         po.fwpartfile = arg
      elif opt in ("-o", "--output"):
@@ -167,10 +210,11 @@ def main(argv):
         po.section_pos[arg_m.group("name")] = int(arg_m.group("pos"),0)
      elif opt in ("-e", "--mkelf"):
         po.command = 'e'
+  po.basename = os.path.splitext(os.path.basename(po.fwpartfile))[0]
   if len(po.fwpartfile) > 0 and len(po.inifile) == 0:
-      po.inifile = os.path.splitext(os.path.basename(po.fwpartfile))[0] + ".a9h"
+      po.inifile = po.basename + ".a9h"
   if len(po.fwpartfile) > 0 and len(po.outfile) == 0:
-      po.outfile = os.path.splitext(os.path.basename(po.fwpartfile))[0] + ".elf"
+      po.outfile = po.basename + ".elf"
   if (po.command == 'e'):
 
     if (po.verbose > 0):
